requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
ccxt==4.4.95
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
frozenlist==1.7.0
idna==3.10
joblib==1.5.1
multidict==6.6.3
nltk==3.9.1
numpy==2.3.1
pandas==2.3.1
propcache==0.3.2
pycares==4.9.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
